Refactoring - first pass
